create_post {distill}	R Documentation

Create a new blog post

Description

Create a new blog post

Usage

create_post(title, author = "auto", slug = "auto",
  date_prefix = TRUE, draft = FALSE, edit = interactive())

Arguments

title
Post title

author
Post author. Automatically drawn from previous post if not provided.

slug
Post slug (directory name). Automatically computed from title if not provided.

date_prefix
Data prefix for post slug (preserves chronological order for posts within the filesystem).

draft
Mark the post as a draft (don't include it in the article listing).

edit
Open the post in an editor after creating it.

Note

This function must be called from with a working directory that is within a Distill website.

Examples

## Not run: 
library(distill)
create_post("My Post")

## End(Not run)

[Package distill version 0.7 Index]