create_website {distill}	R Documentation

Create a Distill website

Description

Create a basic skeleton for a Distill website or blog. Use the create_website() function for a website and the create_blog() function for a blog.

Usage

create_website(dir, title, gh_pages = FALSE, edit = interactive())

create_blog(dir, title, gh_pages = FALSE, edit = interactive())

Arguments

dir
Directory for website

title
Title of website

gh_pages
Configure the site for publishing using GitHub Pages

edit
Open site index file or welcome post in an editor.

Note

The dir and title parameters are required (they will be prompted for interatively if they are not specified).

Examples

## Not run: 
library(distill)
create_website("mysite", "My Site")

## End(Not run)

[Package distill version 0.7 Index]