Publish a Distill Website

Publish a website to RStudio Connect

publish_website(
  site_dir = ".",
  site_name = NULL,
  method = c("rsconnect"),
  server = NULL,
  account = NULL,
  render = TRUE,
  launch_browser = interactive()
)

Arguments

site_dir
Directory containing website. Defaults to current working directory.

site_name
Name for the site (names must be unique within an account). Defaults to the name provided in _site.yml (or to the name of the site_dir if there is no name specified).

method
Publishing method (currently only "rsconnect" is available)

server
Server name. Required only if you use the same account name on multiple servers.

account
Account to deploy application to. This parameter is only required for the initial deployment of an application when there are multiple accounts configured on the system.

render
TRUE to render the site locally before publishing.

launch_browser
If TRUE, the system's default web browser will be launched automatically after the site is deployed. Defaults to TRUE in interactive sessions only.

Examples

if (FALSE) {
library(distill)
publish_website()
}