distill_article {distill}	R Documentation

R Markdown format for Distill articles

Description

Scientific and technical writing, native to the web.

Usage

distill_article(toc = FALSE, toc_depth = 3, fig_width = 6.5,
  fig_height = 4, fig_retina = 2, fig_caption = TRUE, dev = "png",
  smart = TRUE, self_contained = TRUE, mathjax = "default",
  extra_dependencies = NULL, css = NULL, includes = NULL,
  keep_md = FALSE, lib_dir = NULL, md_extensions = NULL,
  pandoc_args = NULL, ...)

Arguments

toc
TRUE to include a table of contents in the output

toc_depth
Depth of headers to include in table of contents

fig_width
Default width (in inches) for figures

fig_height
Default width (in inches) for figures

fig_retina
Scaling to perform for retina displays (defaults to 2, which currently works for all widely used retina displays). Set to NULL to prevent retina scaling. Note that this will always be NULL when keep_md is specified (this is because fig_retina relies on outputting HTML directly into the markdown document).

fig_caption
TRUE to render figures with captions

dev
Graphics device to use for figure output (defaults to png)

smart
Produce typographically correct output, converting straight quotes to curly quotes, — to em-dashes, – to en-dashes, and ... to ellipses.

self_contained
Produce a standalone HTML file with no external dependencies, using data: URIs to incorporate the contents of linked scripts, stylesheets, images, and videos. Note that even for self contained documents MathJax is still loaded externally (this is necessary because of its size).

mathjax
Include mathjax. The "default" option uses an https URL from a MathJax CDN. The "local" option uses a local version of MathJax (which is copied into the output directory). You can pass an alternate URL or pass NULL to exclude MathJax entirely.

extra_dependencies
Additional function arguments to pass to the base R Markdown HTML output formatter html_document_base

css
One or more css files to include

includes
Named list of additional content to include within the document (typically created using the includes function).

keep_md
Keep the markdown file generated by knitting.

lib_dir
Directory to copy dependent HTML libraries (e.g. jquery, bootstrap, etc.) into. By default this will be the name of the document with _files appended to it.

md_extensions
Markdown extensions to be added or removed from the default definition or R Markdown. See the rmarkdown_format for additional details.

pandoc_args
Additional command line options to pass to pandoc

...
Additional function arguments to pass to the base R Markdown HTML output formatter html_document_base

Details

Distill articles feature attractive, reader-friendly typography, flexible layout options for visualizations, and full support for footnotes and citations.

[Package distill version 0.7 Index]